+++ src/algorithm/icla_algorithm.c
@@ -171,7 +171,7 @@ static inline void computeMovingDirection(IcmPoint *points, int count) {
 static inline void
 __attribute__((__always_inline__,__gnu_inline__,__artificial__,__nonnull__))
 icla_run(const VECTOR* vx, const VECTOR* vy, const VECTOR *restrict r,
-              size_t no_anchors, int width, int height,
+              size_t no_anchors, size_t width, size_t height,
               VECTOR *restrict resx, VECTOR *restrict resy)
 {
     static const double alpha = 1.5;
