+++ src/algorithm/minmax_algorithm.c
@@ -41,8 +41,8 @@ static inline void
 __attribute__((__always_inline__,__gnu_inline__,__artificial__,__nonnull__))
 minmax_run (const VECTOR *restrict vx, const VECTOR *restrict vy,
             const VECTOR *restrict r, size_t num_anchors,
-            int width __attribute__((__unused__)),
-            int height __attribute__((__unused__)),
+            size_t width __attribute__((__unused__)),
+            size_t height __attribute__((__unused__)),
             VECTOR *restrict resx, VECTOR *restrict resy)
 {
     VECTOR east  = vx[0] + r[0];
