+++ src/algorithm/mle_gauss_algorithm.c
@@ -193,7 +193,7 @@ mle_gauss_likelihood_fdf(const gsl_vector *X, void *restrict params,
 static inline void
 __attribute__((__always_inline__,__gnu_inline__,__artificial__,__nonnull__))
 mle_gauss_run(const VECTOR* vx, const VECTOR* vy, const VECTOR *restrict r,
-              size_t no_anchors, int width, int height,
+              size_t no_anchors, size_t width, size_t height,
               VECTOR *restrict resx, VECTOR *restrict resy)
 {
     /* Step 0: Set up the likelihood function. */
